add one off pod to generate access.json (#3)
* add one off pod to generate access.json
* add updated testing container

This still will not work because it does not register using the access.json,
and it seems to still want write. I am concerned this will not work without
a RWX volume, which is not ideal.

this is one step closer in that we can generate the config
in advance. What we still need to be able to do is start
a server (vs. a worker) on separate hosts. Right now the config
only accepts one hostname

* everything is working!

We need to wait for this new feature to be merged (and then update
our example and recipes here) but this is working very nicely

* tweak readme to remove spelling error
* separate lammps from hello-world and install nightly release of hq
* update with jobsetv2 release and hyperqueue release

Signed-off-by: vsoch <vsoch@users.noreply.github.com>